Seal ProtoDecoder
Merged ProtoDecoderImpl with ProtoDecoder, allowing for the use of the sealed keyword. This makes more sense now, the only downside is that the resulting file is pretty long... but I think we can live with that.
