+++ core/src/main/scala/eu/ostrzyciel/jelly/core/ProtoDecoder.scala
@@ -4,16 +4,15 @@ import eu.ostrzyciel.jelly.core.proto.v1.*
 import eu.ostrzyciel.jelly.core.proto.v1.RdfGraph.Graph
 
 import scala.annotation.tailrec
+import scala.collection.mutable.ListBuffer
 import scala.reflect.ClassTag
 
 /**
  * Base class for stateful decoders of protobuf RDF streams.
  *
  * See the implementations in [[ProtoDecoderImpl]].
- *
- * @note This trait is not stable for extension.
  */
-abstract class ProtoDecoder[TNode, TDatatype : ClassTag, TTriple, TQuad, TOut]
+sealed abstract class ProtoDecoder[TNode, TDatatype : ClassTag, TTriple, TQuad, TOut]
 (converter: ProtoDecoderConverter[TNode, TDatatype, TTriple, TQuad]):
   private var streamOpt: Option[RdfStreamOptions] = None
   private lazy val nameDecoder = new NameDecoder(streamOpt getOrElse RdfStreamOptions.defaultInstance)
@@ -149,3 +148,108 @@ abstract class ProtoDecoder[TNode, TDatatype : ClassTag, TTriple, TQuad, TOut]
 
   protected def handleGraphEnd(): Option[TOut] =
     throw new RdfProtoDeserializationError("Unexpected end of graph in stream.")
+
+
+/**
+ * Implementations of the [[ProtoDecoder]] trait.
+ */
+object ProtoDecoder:
+
+  /**
+   * A decoder that reads TRIPLES streams and outputs a sequence of triples.
+   */
+  final class TriplesDecoder[TNode, TDatatype : ClassTag, TTriple, TQuad]
+  (converter: ProtoDecoderConverter[TNode, TDatatype, TTriple, TQuad])
+    extends ProtoDecoder[TNode, TDatatype, TTriple, TQuad, TTriple](converter):
+
+    override protected def handleOptions(opts: RdfStreamOptions): Unit =
+      if !opts.streamType.isRdfStreamTypeTriples then
+        throw new RdfProtoDeserializationError("Incoming stream type is not TRIPLES.")
+      super.handleOptions(opts)
+
+    override protected def handleTriple(triple: RdfTriple): Option[TTriple] =
+      Some(convertTriple(triple))
+
+  /**
+   * A decoder that reads QUADS streams and outputs a sequence of quads.
+   */
+  final class QuadsDecoder[TNode, TDatatype : ClassTag, TTriple, TQuad]
+  (converter: ProtoDecoderConverter[TNode, TDatatype, TTriple, TQuad])
+    extends ProtoDecoder[TNode, TDatatype, TTriple, TQuad, TQuad](converter):
+
+    override protected def handleOptions(opts: RdfStreamOptions): Unit =
+      if !opts.streamType.isRdfStreamTypeQuads then
+        throw new RdfProtoDeserializationError("Incoming stream type is not QUADS.")
+      super.handleOptions(opts)
+
+    override protected def handleQuad(quad: RdfQuad): Option[TQuad] =
+      Some(convertQuad(quad))
+
+  /**
+   * A decoder that reads GRAPHS streams and outputs a flat sequence of quads.
+   */
+  final class GraphsAsQuadsDecoder[TNode, TDatatype : ClassTag, TTriple, TQuad]
+  (converter: ProtoDecoderConverter[TNode, TDatatype, TTriple, TQuad])
+    extends ProtoDecoder[TNode, TDatatype, TTriple, TQuad, TQuad](converter):
+    private var currentGraph: Option[TNode] = None
+
+    override protected def handleOptions(opts: RdfStreamOptions): Unit =
+      if !opts.streamType.isRdfStreamTypeGraphs then
+        throw new RdfProtoDeserializationError("Incoming stream type is not GRAPHS.")
+      super.handleOptions(opts)
+
+    override protected def handleGraphStart(graph: RdfGraphStart): Option[TQuad] =
+      currentGraph = Some(convertGraphTerm(graph.graph))
+      None
+
+    override protected def handleGraphEnd(): Option[TQuad] =
+      currentGraph = None
+      None
+
+    override protected def handleTriple(triple: RdfTriple): Option[TQuad] =
+      if currentGraph.isEmpty then
+        throw new RdfProtoDeserializationError("Triple in stream without preceding graph start.")
+      Some(converter.makeQuad(
+        convertTermWrapped(triple.s, lastSubject),
+        convertTermWrapped(triple.p, lastPredicate),
+        convertTermWrapped(triple.o, lastObject),
+        currentGraph.get,
+      ))
+
+  /**
+   * A decoder that reads GRAPHS streams and outputs a sequence of graphs.
+   * Each graph is emitted as soon as the producer signals that it's complete.
+   */
+  final class GraphsDecoder[TNode, TDatatype : ClassTag, TTriple, TQuad]
+  (converter: ProtoDecoderConverter[TNode, TDatatype, TTriple, TQuad])
+    extends ProtoDecoder[TNode, TDatatype, TTriple, TQuad, (TNode, Iterable[TTriple])](converter):
+    private var currentGraph: Option[TNode] = None
+    private var buffer: ListBuffer[TTriple] = new ListBuffer[TTriple]()
+
+    override protected def handleOptions(opts: RdfStreamOptions): Unit =
+      if !opts.streamType.isRdfStreamTypeGraphs then
+        throw new RdfProtoDeserializationError("Incoming stream type is not GRAPHS.")
+      super.handleOptions(opts)
+
+    private inline def emitBuffer(): Option[(TNode, Iterable[TTriple])] =
+      if buffer.isEmpty then None
+      else if currentGraph.isEmpty then
+        throw new RdfProtoDeserializationError("End of graph encountered before a start.")
+      else
+        Some((currentGraph.get, buffer))
+
+    override protected def handleGraphStart(graph: RdfGraphStart): Option[(TNode, Iterable[TTriple])] =
+      val toEmit = emitBuffer()
+      buffer = new ListBuffer[TTriple]()
+      currentGraph = Some(convertGraphTerm(graph.graph))
+      toEmit
+
+    override protected def handleGraphEnd(): Option[(TNode, Iterable[TTriple])] =
+      val toEmit = emitBuffer()
+      buffer = new ListBuffer[TTriple]()
+      currentGraph = None
+      toEmit
+
+    override protected def handleTriple(triple: RdfTriple): Option[(TNode, Iterable[TTriple])] =
+      buffer.addOne(convertTriple(triple))
+      None
