--- core/src/main/scala/eu/ostrzyciel/jelly/core/ProtoDecoderImpl.scala
@@ -1,110 +1,0 @@
-package eu.ostrzyciel.jelly.core
-
-import eu.ostrzyciel.jelly.core.proto.v1.*
-
-import scala.collection.mutable.{ArrayBuffer, ListBuffer}
-import scala.reflect.ClassTag
-
-/**
- * Implementations of the [[ProtoDecoder]] trait.
- */
-object ProtoDecoderImpl:
-
-  /**
-   * A decoder that reads TRIPLES streams and outputs a sequence of triples.
-   */
-  final class TriplesDecoder[TNode, TDatatype : ClassTag, TTriple, TQuad]
-  (converter: ProtoDecoderConverter[TNode, TDatatype, TTriple, TQuad])
-    extends ProtoDecoder[TNode, TDatatype, TTriple, TQuad, TTriple](converter):
-
-    override protected def handleOptions(opts: RdfStreamOptions): Unit =
-      if !opts.streamType.isRdfStreamTypeTriples then
-        throw new RdfProtoDeserializationError("Incoming stream type is not TRIPLES.")
-      super.handleOptions(opts)
-
-    override protected def handleTriple(triple: RdfTriple): Option[TTriple] =
-      Some(convertTriple(triple))
-
-  /**
-   * A decoder that reads QUADS streams and outputs a sequence of quads.
-   */
-  final class QuadsDecoder[TNode, TDatatype : ClassTag, TTriple, TQuad]
-  (converter: ProtoDecoderConverter[TNode, TDatatype, TTriple, TQuad])
-    extends ProtoDecoder[TNode, TDatatype, TTriple, TQuad, TQuad](converter):
-
-    override protected def handleOptions(opts: RdfStreamOptions): Unit =
-      if !opts.streamType.isRdfStreamTypeQuads then
-        throw new RdfProtoDeserializationError("Incoming stream type is not QUADS.")
-      super.handleOptions(opts)
-
-    override protected def handleQuad(quad: RdfQuad): Option[TQuad] =
-      Some(convertQuad(quad))
-
-  /**
-   * A decoder that reads GRAPHS streams and outputs a flat sequence of quads.
-   */
-  final class GraphsAsQuadsDecoder[TNode, TDatatype : ClassTag, TTriple, TQuad]
-  (converter: ProtoDecoderConverter[TNode, TDatatype, TTriple, TQuad])
-    extends ProtoDecoder[TNode, TDatatype, TTriple, TQuad, TQuad](converter):
-    private var currentGraph: Option[TNode] = None
-
-    override protected def handleOptions(opts: RdfStreamOptions): Unit =
-      if !opts.streamType.isRdfStreamTypeGraphs then
-        throw new RdfProtoDeserializationError("Incoming stream type is not GRAPHS.")
-      super.handleOptions(opts)
-
-    override protected def handleGraphStart(graph: RdfGraphStart): Option[TQuad] =
-      currentGraph = Some(convertGraphTerm(graph.graph))
-      None
-
-    override protected def handleGraphEnd(): Option[TQuad] =
-      currentGraph = None
-      None
-
-    override protected def handleTriple(triple: RdfTriple): Option[TQuad] =
-      if currentGraph.isEmpty then
-        throw new RdfProtoDeserializationError("Triple in stream without preceding graph start.")
-      Some(converter.makeQuad(
-        convertTermWrapped(triple.s, lastSubject),
-        convertTermWrapped(triple.p, lastPredicate),
-        convertTermWrapped(triple.o, lastObject),
-        currentGraph.get,
-      ))
-
-  /**
-   * A decoder that reads GRAPHS streams and outputs a sequence of graphs.
-   * Each graph is emitted as soon as the producer signals that it's complete.
-   */
-  final class GraphsDecoder[TNode, TDatatype : ClassTag, TTriple, TQuad]
-  (converter: ProtoDecoderConverter[TNode, TDatatype, TTriple, TQuad])
-    extends ProtoDecoder[TNode, TDatatype, TTriple, TQuad, (TNode, Iterable[TTriple])](converter):
-    private var currentGraph: Option[TNode] = None
-    private var buffer: ListBuffer[TTriple] = new ListBuffer[TTriple]()
-
-    override protected def handleOptions(opts: RdfStreamOptions): Unit =
-      if !opts.streamType.isRdfStreamTypeGraphs then
-        throw new RdfProtoDeserializationError("Incoming stream type is not GRAPHS.")
-      super.handleOptions(opts)
-
-    private inline def emitBuffer(): Option[(TNode, Iterable[TTriple])] =
-      if buffer.isEmpty then None
-      else if currentGraph.isEmpty then
-        throw new RdfProtoDeserializationError("End of graph encountered before a start.")
-      else
-        Some((currentGraph.get, buffer))
-
-    override protected def handleGraphStart(graph: RdfGraphStart): Option[(TNode, Iterable[TTriple])] =
-      val toEmit = emitBuffer()
-      buffer = new ListBuffer[TTriple]()
-      currentGraph = Some(convertGraphTerm(graph.graph))
-      toEmit
-
-    override protected def handleGraphEnd(): Option[(TNode, Iterable[TTriple])] =
-      val toEmit = emitBuffer()
-      buffer = new ListBuffer[TTriple]()
-      currentGraph = None
-      toEmit
-
-    override protected def handleTriple(triple: RdfTriple): Option[(TNode, Iterable[TTriple])] =
-      buffer.addOne(convertTriple(triple))
-      None
